refactor(cli): split ui.py into a cli/ui/ package
The single 1344-line ui.py mixed branding, repo pre-scan, .env persistence,
provider/mode selection, result panels, workspace selection, and the Rich
progress adapter. Split it into a cli/ui/ package, one module per concern,
with a façade __init__ that re-exports every previously-public name so all
existing 'from repowise.cli.ui import ...' call sites are unchanged.

interactive_advanced_config is decomposed into per-section helpers
(_prompt_scope / _prompt_run_mode / _prompt_exclude / _prompt_git /
_prompt_generation / _build_summary_table) — same prompts, same order,
same returned dict.

Pure mechanical move; no behavior change. Adds tests/unit/cli/test_ui_package.py
(façade parity + advanced-config characterization).
